+++ lib/calendar/getInfo.js
