+++ lib/calendar/renderCalendar.js
