+++ lib/index.js
@@ -6,7 +6,7 @@ export * from './error/index.js'
 export * from './utils/index.js'
 export * from './animation/index.js'
 export * from './son/index.js'
-export * from './kim/index.js'
+export * from './calendar/index.js'
 
 export * from './storage/index.js'
 export * from './note/index.js'
