+++ lib/jung/handleBookmarkBtnClick.js
@@ -1,6 +1,7 @@
 import { getMyBookmarks, addBookmark, deleteBookmark } from '../api/index.js'
 import { getAttr } from '../dom/attr.js'
 import { getUserSessionStorage } from '../storage/sessionStorage.js'
+import { renderCalendar } from '../index.js'
 
 function getPostId(e) {
   if (e.target.tagName === 'BUTTON') {
@@ -30,10 +31,12 @@ export function handleBookmarkBtnClick(e) {
         addBookmark(currentUser, id)
         target.classList.add('activate')
         target.textContent = '내 관심 공고'
+        renderCalendar();
       } else if (data.bookmarklist.includes(id)) {
         deleteBookmark(currentUser, id)
         target.classList.remove('activate')
         target.textContent = '관심목록 +'
+        renderCalendar();
       }
     })
 }
